+++ src/features/pathCompletions.ts
@@ -63,6 +63,13 @@ interface CompletionContext {
 	readonly anchorInfo?: AnchorContext;
 }
 
+function tryDecodeUriComponent(str: string): string {
+	try {
+		return decodeURIComponent(str);
+	} catch {
+		return str;
+	}
+}
 export class PathCompletionProvider implements vscode.CompletionItemProvider {
 
 	public static register(selector: vscode.DocumentSelector, engine: TextileEngine): vscode.Disposable {
@@ -90,7 +97,7 @@ export class PathCompletionProvider implements vscode.CompletionItemProvider {
 
 			case CompletionContextKind.LinkDefinition:
 			case CompletionContextKind.Link: {
-				const items: vscode.CompletionItem[] = Array.from(this.provideReferenceSuggestions(document, position, context));
+				const items: vscode.CompletionItem[] = Array.from(this.provideReferenceSuggestions(document, position, context)); // Modified for Textile
 
 				const isAnchorInCurrentDoc = context.anchorInfo && context.anchorInfo.beforeAnchor.length === 0;
 
@@ -162,7 +169,7 @@ export class PathCompletionProvider implements vscode.CompletionItemProvider {
 			const suffix = lineSuffixText.match(/^[^\s]*/); // Changed for Textile
 			return {
 				kind: CompletionContextKind.Link,
-				linkPrefix: prefix,
+				linkPrefix: tryDecodeUriComponent(prefix),
 				linkTextStartPosition: position.translate({ characterDelta: -prefix.length }),
 				linkSuffix: suffix ? suffix[0] : '',
 				anchorInfo: this.getAnchorContext(prefix),
@@ -179,7 +186,7 @@ export class PathCompletionProvider implements vscode.CompletionItemProvider {
 			const suffix = lineSuffixText.match(/^[^\s]*/);
 			return {
 				kind: CompletionContextKind.LinkDefinition,
-				linkPrefix: prefix,
+				linkPrefix: tryDecodeUriComponent(prefix),
 				linkTextStartPosition: position.translate({ characterDelta: -prefix.length }),
 				linkSuffix: suffix ? suffix[0] : '',
 				anchorInfo: this.getAnchorContext(prefix),