+++ src/features/preview.ts
@@ -5,13 +5,13 @@
 
 import * as vscode from 'vscode';
 import * as nls from 'vscode-nls';
+import * as uri from 'vscode-uri';
 import { Logger } from '../logger';
 import { TextileEngine } from '../textileEngine';
 import { TextileContributionProvider } from '../textileExtensions';
 import { Disposable } from '../util/dispose';
 import { isTextileFile } from '../util/file';
 import { openDocumentLink, resolveDocumentLink, resolveUriToTextileFile } from '../util/openDocumentLink';
-import * as path from '../util/path';
 import { WebviewResourceProvider } from '../util/resources';
 import { getVisibleLine, LastScrollLocation, TopmostLineMonitor } from '../util/topmostLineMonitor';
 import { urlToUri } from '../util/url';
@@ -177,31 +177,6 @@ class TextilePreview extends Disposable implements WebviewResourceProvider {
 			}
 		}));
 
-		this._register(this._webviewPanel.onDidChangeViewState(async () => {
-			if (this._disposed) {
-				return;
-			}
-
-			if (this._webviewPanel.active) {
-				let document: vscode.TextDocument;
-				try {
-					document = await vscode.workspace.openTextDocument(this._resource);
-				} catch {
-					return;
-				}
-
-				if (this._disposed) {
-					return;
-				}
-
-				const content = await this._contentProvider.provideTextDocumentContent(document, this, this._previewConfigurations, this.line, this.state);
-				if (!this._webviewPanel.active && !this._disposed) {
-					// Update the html so we can show it properly when restoring it
-					this._webviewPanel.webview.html = content.html;
-				}
-			}
-		}));
-
 		this._register(this._webviewPanel.webview.onDidReceiveMessage((e: CacheImageSizesMessage | RevealLineMessage | DidClickMessage | ClickLinkMessage | ShowPreviewSecuritySelectorMessage | PreviewStyleLoadErrorMessage) => {
 			if (e.source !== this._resource.toString()) {
 				return;
@@ -339,7 +314,7 @@ class TextilePreview extends Disposable implements WebviewResourceProvider {
 			return;
 		}
 
-		const shouldReloadPage = forceUpdate || !this.currentVersion || this.currentVersion.resource.toString() !== pendingVersion.resource.toString();
+		const shouldReloadPage = forceUpdate || !this.currentVersion || this.currentVersion.resource.toString() !== pendingVersion.resource.toString() || !this._webviewPanel.visible;
 		this.currentVersion = pendingVersion;
 		const content = await (shouldReloadPage
 			? this._contentProvider.provideTextDocumentContent(document, this, this._previewConfigurations, this.line, this.state)
@@ -421,7 +396,7 @@ class TextilePreview extends Disposable implements WebviewResourceProvider {
 		const srcs = new Set(containingImages.map(img => img.src));
 
 		// Delete stale file watchers.
-		for (const [src, watcher] of [...this._fileWatchersBySrc]) {
+		for (const [src, watcher] of this._fileWatchersBySrc) {
 			if (!srcs.has(src)) {
 				watcher.dispose();
 				this._fileWatchersBySrc.delete(src);
@@ -460,7 +435,7 @@ class TextilePreview extends Disposable implements WebviewResourceProvider {
 				baseRoots.push(...workspaceRoots);
 			}
 		} else {
-			baseRoots.push(this._resource.with({ path: path.dirname(this._resource.path) }));
+			baseRoots.push(uri.Utils.dirname(this._resource));
 		}
 
 		return baseRoots;
@@ -786,9 +761,10 @@ export class DynamicTextilePreview extends Disposable implements ManagedTextileP
 	}
 
 	private static getPreviewTitle(resource: vscode.Uri, locked: boolean): string {
+		const resourceLabel = uri.Utils.basename(resource);
 		return locked
-			? localize('lockedPreviewTitle', '[Preview] {0}', path.basename(resource.fsPath))
-			: localize('previewTitle', 'Preview {0}', path.basename(resource.fsPath));
+			? localize('lockedPreviewTitle', '[Preview] {0}', resourceLabel)
+			: localize('previewTitle', 'Preview {0}', resourceLabel);
 	}
 
 	public get position(): vscode.ViewColumn | undefined {