+++ src/features/previewContentProvider.ts
@@ -5,11 +5,11 @@
 
 import * as vscode from 'vscode';
 import * as nls from 'vscode-nls';
+import * as uri from 'vscode-uri';
 import { Logger } from '../logger';
 import { TextileEngine } from '../textileEngine';
 import { TextileContributionProvider } from '../textileExtensions';
 import { ContentSecurityPolicyArbiter, TextilePreviewSecurityLevel } from '../security';
-import { basename, dirname, isAbsolute, join } from '../util/path';
 import { WebviewResourceProvider } from '../util/resources';
 import { TextilePreviewConfiguration, TextilePreviewConfigurationManager } from './previewConfig';
 
@@ -127,7 +127,7 @@ export class TextileContentProvider {
 	public provideFileNotFoundContent(
 		resource: vscode.Uri,
 	): string {
-		const resourcePath = basename(resource.fsPath);
+		const resourcePath = uri.Utils.basename(resource);
 		const body = localize('preview.notFound', '{0} cannot be found', resourcePath);
 		return `<!DOCTYPE html>
 			<html>
@@ -153,7 +153,7 @@ export class TextileContentProvider {
 		}
 
 		// Assume it must be a local file
-		if (isAbsolute(href)) {
+		if (href.startsWith('/') || /^[a-z]:\\/i.test(href)) {
 			return resourceProvider.asWebviewUri(vscode.Uri.file(href)).toString();
 		}
 
@@ -164,7 +164,7 @@ export class TextileContentProvider {
 		}
 
 		// Otherwise look relative to the textile file
-		return resourceProvider.asWebviewUri(vscode.Uri.file(join(dirname(resource.fsPath), href))).toString();
+		return resourceProvider.asWebviewUri(vscode.Uri.joinPath(uri.Utils.dirname(resource), href)).toString();
 	}
 
 	private computeCustomStyleSheetIncludes(resourceProvider: WebviewResourceProvider, resource: vscode.Uri, config: TextilePreviewConfiguration): string {