+++ src/test/foldingProvider.test.ts
@@ -4,12 +4,11 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as assert from 'assert';
-import * as vscode from 'vscode';
 import 'mocha';
-
+import * as vscode from 'vscode';
 import TextileFoldingProvider from '../features/foldingProvider';
-import { InMemoryDocument } from './inMemoryDocument';
 import { createNewTextileEngine } from './engine';
+import { InMemoryDocument } from './inMemoryDocument';
 
 const testFileName = vscode.Uri.file('test.md');